Changed the type of the buffer length to size_t
Why was this an int in the first place?

+++ decoder/packet.h
@@ -138,7 +138,7 @@ void clear_statistics();
 uint16_t get_packet_size(struct packet *p);
 
 /* Should be internal to the decoder */
-packet_error check_port(unsigned char *buffer, int n,
+packet_error check_port(unsigned char *buffer, size_t n,
                         struct application_info *adu, uint16_t port);
 
 #endif
